+++ test/test_ads_fp_gen.py
@@ -13,9 +13,13 @@ from catlearn.api.ase_atoms_api import database_to_list, images_connectivity
 from catlearn.fingerprint.adsorbate_prep import (autogen_info,
                                                  check_reconstructions,
                                                  connectivity2ads_index,
-                                                 slab_positions2ads_index)
+                                                 slab_positions2ads_index,
+                                                 slab_index,
+                                                 attach_cations,
+                                                 info2primary_index)
 from catlearn.fingerprint.periodic_table_data import (get_radius,
-                                                      default_catlearn_radius)
+                                                      default_catlearn_radius,
+                                                      stat_mendeleev_params)
 from catlearn.fingerprint.setup import FeatureGenerator, default_fingerprinters
 
 wkdir = os.getcwd()
@@ -24,7 +28,7 @@ wkdir = os.getcwd()
 class TestAdsorbateFeatures(unittest.TestCase):
     """Test out the adsorbate feature generation."""
 
-    def setup_atoms(self):
+    def setup_metals(self):
         """Get the atoms objects."""
         adsorbates = ['H', 'O', 'C', 'N', 'S', 'Cl', 'P', 'F']
         symbols = ['Ag', 'Au', 'Cu', 'Pt', 'Pd', 'Ir', 'Rh', 'Ni', 'Co']
@@ -43,7 +47,7 @@ class TestAdsorbateFeatures(unittest.TestCase):
 
     def test_tags(self):
         """Test the feature generation."""
-        images = self.setup_atoms()
+        images = self.setup_metals()
         images = autogen_info(images)
         print(str(len(images)) + ' training examples.')
         gen = FeatureGenerator(nprocs=1)
@@ -61,7 +65,7 @@ class TestAdsorbateFeatures(unittest.TestCase):
 
     def test_constrained_ads(self):
         """Test the feature generation."""
-        images = self.setup_atoms()
+        images = self.setup_metals()
         [atoms.set_tags(np.zeros(len(atoms))) for atoms in images]
         for atoms in images:
             c_atoms = [a.index for a in atoms if
@@ -112,9 +116,42 @@ class TestAdsorbateFeatures(unittest.TestCase):
         reconstructed = check_reconstructions(image_pairs)
         for i in range(len(images)):
             species = images[i].info['key_value_pairs']['species']
-            slab_positions2ads_index(images[i], slabs[i], species)
             connectivity2ads_index(images[i], species)
         self.assertTrue(len(reconstructed) == 0)
+
+    def test_chalcogenides(self):
+        images = database_to_list('data/bajdichWO32018_ads.db')
+        images = images_connectivity(images)
+        slabs = database_to_list('data/bajdichWO32018_slabs.db')
+        slabs_dict = {}
+        for slab in slabs:
+            slabs_dict[slab.info['id']] = slab
+        for i in range(len(images)):
+            species = images[i].info['key_value_pairs']['species']
+            images[i].subsets['ads_atoms'] = \
+                slab_positions2ads_index(images[i], slabs[i], species)
+            if 'slab_atoms' not in images[i].subsets:
+                images[i].subsets['slab_atoms'] = slab_index(images[i])
+            if ('chemisorbed_atoms' not in images[i].subsets or
+                'site_atoms' not in images[i].subsets or
+                    'ligand_atoms' not in images[i].subsets):
+                chemi, site, ligand = info2primary_index(images[i])
+                images[i].subsets['chemisorbed_atoms'] = chemi
+                images[i].subsets['site_atoms'] = site
+                images[i].subsets['ligand_atoms'] = ligand
+            attach_cations(images[i], anion_number=8)
+        gen = FeatureGenerator(nprocs=1)
+        train_fpv = default_fingerprinters(gen, 'chalcogenides')
+        matrix = gen.return_vec(images, train_fpv)
+        labels = gen.return_names(train_fpv)
+        if __name__ == '__main__':
+            for i, l in enumerate(labels):
+                print(i, l)
+        self.assertTrue(len(labels) == np.shape(matrix)[1])
+
+    def test_periodic_table(self):
+        r, w = stat_mendeleev_params('MoS2', params=None)
+        self.assertTrue(len(w) == np.shape(r)[0])
 
 
 if __name__ == '__main__':